type

+++ DoiFApp/Data/Models/LessonModel.cs
@@ -14,6 +14,6 @@
         public List<string> Groups { get; set; } = [];
         public List<string> Auditoriums { get; set; } = [];
 
-        public int Wight { get; set; } = 2;
+        public double Wight { get; set; } = 2;
     }
 }
